Host.cpp: generate alerts for local hosts only
This commit lets alerts be generated only for local hosts and not
for remote hosts. When a remote host can be mentioned in an alert
because it holds a relevant role in it (e.g. it is an attacker
of a local host) it is mentioned but not specific alert is generated
for it. This addresses issue #3.