src/DB.cpp: fix issue with flow dumps
This commit changes the DB handling backend so that it correctly
dumps all kind of strings. In fact, some of the strings (e.g. the
HTTP URL in the flow JSON) could contain single or double quotes,
which could result in DB-related errors. To correctly handle all
kind of input strings, a prepared statement is built (escaping
the string would leave room to unhandled scenarios).