src/Prefs.cpp: add parameter to specify install directory
Starting ntopng from outside the install folder can cause problems
if ntopng has not been installed but instead compiled from source.
This commit adds an install-dir parameter to allow to specify where
ntopng executable and scripts are so that variables are set
correctly. Please refer to the help to get info about the usage
of the option.