src: add communities manager
This commit introduces a class to handle host communities, that are
a logical representation of clusters of hosts that fall under the
same administrative domain. This commit also adds a new command line
option to specify a file containing aggregation rules and code to
parse such file with an adequate syntax:
  communityX@id1=net1,net2,net3
  communityY@id2=net4,net5,net6
  ...

Communities are unused now. Further commits will introduce reports
to represent communities and describe them with stats.
This addresses issue #75.