Implemented dissectSSL
Studing the ssl handshake, i understand to calculate the ssl handshake time. First ssl packet is client-hello (with code 1), the last is new session ticket (with code 4), the delta time between these 2 packet is the ssl hs time. I have to fix the granularity of the time, next step is to see how much time pass between the end of the handshake and the first data packet, and from the first data packet and the reply (another data packet, if the server send an ack, the incoming data packet was not complete)