Implements counter resets
This commits adds two reset buttons to the interface page.

One button allows the reset of all interface counters, namely,
packets and bytes received, packets dropped, and the number of
flows exported and dropped (when mysql or elasticsearch is enabled).

The other button allows the reset of only the drops.