Security fixes (escapes and double-decode)
Prevents single (') and double (") quotes to be input as they may be used to inject arbitrary payload in requests.

Prevents double-decodes of url-encoded http contents as they can be exploited to inject non-allowed characters
in multiple steps (e.g., %2522 after the first decode becomes %22 and after the second decode becomes ").