Enforces shapers and quotas for user-defined pools
Shapers and quotas are no longer allowed in the default (not assigned) host pool.
This means that the default host pool has become pass-all.

The idea behind this choice is that a user should enforce policies only on the members
for custom pools. Before this commit, a quota or a shaper in the default (not assigned)
host pool, could interfere with a quota or a shaper in a custom host pool.