Fix pcap packet poll loop hang when no packet is received.
As stated in https://www.tcpdump.org/manpages/pcap_breakloop.3pcap.html
pcap_breakloop should be called on the same blocking thread in order to succeed.
A term signal is now sent to the packet loop thread to break the loop.

Here is a backtrace of the blocked polling thread.

Thread 15 (Thread 0x7fcfe17f2700 (LWP 9568)):