Fixes false SSL cert mismatch due to wrong case-sensitive checks
This was happening for example when requesting

https://images-na.ssl-images-amazon.com/

As the response from the server was for (note the cap I):
Images-na.ssl-images-amazon.com

CNs are NOT case sensitive as per RFC 5280:

When the subjectAltName extension contains a domain name system
label, the domain name MUST be stored in the dNSName (an IA5String).
The name MUST be in the "preferred name syntax", as specified by
Section 3.5 of RFC1034 and as modified by Section 2.1 of
RFC1123. Note that while uppercase and lowercase letters are
allowed in domain names, no significance is attached to the case.