Upgrade C++ standard to C++1y standard (#6522)
This option works on CentOS 7 (GCC 4.8) and above, other Linux OSes have been upgraded to higher GCC, so CentOS 7 needs to be considered, C++1y (before C++14) is what it can do.
The point for advanced versions of C++ is that we can use a lot of libraries. There are currently a lot of libraries that need to be basically C++11.